Moved Header and Footer in components
