+++ src/components/Layout/components/Header/components/index.ts
@@ -1,0 +1,2 @@
+export * from './GitHubButton';
+export * from './NotificationsButton';
